중복된 소셜 리다이렉트 컴포넌트 하나로 통합
